Added JSONType, JSONObjectType, and SchemaEnv.
JSONType is the top level interface for all types expressed in
JSONSchema.  It comes with four static inner classes, expressing the
types of the built-in JSON strings, integers, doubles, and booleans.

JSONObjecType is the beginning of the implementation for parsing
in a schema representation of an object type and validating it.

(Still undone: JSONArrayType).

SchemaEnv is a "schema environment", the set of bindings of names
to types, in a hierarchical structure.