Updated Partial Implementation.
Added a stub for the Array type, and expanded the
SchemaEnv implementation with a new constructor (for a top-level
environment built out of the .js files in a directory) and
with the "eval" method, which is the entry point to the
Schema parser.

SchemaException is the general (checked) exception thrown by the
parser's apparatus.

JSONFileType is a "stub" type, which allows for delayed type
parsing from .js files in the filesystem. (We need this, in order
for mutually-recursive top-level types in a filesystem to work
correctly).