Fixed the reading of schema from a file.
For some reason (?), the new JSONObject(new FileReader(f))
motif wasn't working for creating the JSONObject with which to
initialize the JSONObjectType -- the latter class was seeing a
blank object, {}, in its constructor.  Probably I need to call a
"parse" or something, but instead I've simply cached the file
as a String and then fed that string directly to the JSONObject.
Blah.