Changed how JSONObjectType handles 'extra' keys.
Before, JSONObjectType would signal a validation error, if we
encountered a key in the value which was not present in the schema.
This behavior is too strict -- we want to ensure that *all* the keys
from the schema are present in the value, but extra keys in the value
may indicate a "subtype," and so we don't want to throw an exception
unless there's some additional annotation on the schema (which isn't
supported now, but maybe in the future -- for example, some kind of
'strict' or 'closed-world' validation mode).